Move all mock data for system tests to same place and improve names #302
